Inside the "brute force solutions" source and test folders, rename "FindPairWithSum" with "TwoSum."
